requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
pytz==2018.9
